Feat: 삭제 엔티티 미조회 추가

+++ dearwith-backend/src/main/java/com/dearwith/dearwith_backend/review/dto/ReviewLikeResponseDto.java
@@ -3,5 +3,5 @@ package com.dearwith.dearwith_backend.review.dto;
 public record ReviewLikeResponseDto(
         Long reviewId,
         boolean liked,
-        long likeCount
+        Integer likeCount
 ) {}
