+++ dearwith-backend/src/main/java/com/dearwith/dearwith_backend/review/repository/ReviewRepository.java
@@ -60,5 +60,5 @@ public interface ReviewRepository extends JpaRepository<Review, Long> {
     Optional<Review> findById(Long id);
 
     @Query("select r.likeCount from Review r where r.id = :reviewId")
-    long getLikeCount(@Param("reviewId") Long reviewId);
+    Integer getLikeCount(@Param("reviewId") Long reviewId);
 }
